Make exceptions easier to test
