+++ src/Spectre.Console.Tests/Expectations/Exception/GenericException.Output_exceptionFormats=Default.verified.txt
@@ -1,4 +1,4 @@
 ﻿Spectre.Console.Tests.Data.GenericException<Spectre.Console.IAnsiConsole>: Throwing!
-  at bool Spectre.Console.Tests.Data.TestExceptions.MethodThatThrowsGenericException<T>() in /xyz/Exceptions.cs:nn
-  at void Spectre.Console.Tests.Unit.ExceptionTests.<>c.<Should_Write_GenericException>b__8_0() in /xyz/ExceptionTests.cs:nn
-  at Exception Spectre.Console.Tests.Unit.ExceptionTests.GetException(Action action) in /xyz/ExceptionTests.cs:nn
+  at bool Spectre.Console.Tests.Data.TestExceptions.MethodThatThrowsGenericException<T>() in /xyz/Exceptions.cs:100          
+  at void Spectre.Console.Tests.Unit.ExceptionTests.<>c.<Should_Write_GenericException>b__8_0() in /xyz/ExceptionTests.cs:200
+  at Exception Spectre.Console.Tests.Unit.ExceptionTests.GetException(Action action) in /xyz/ExceptionTests.cs:300
