+++ src/Spectre.Console.Tests/Expectations/Exception/GenericException.Output_exceptionFormats=ShortenMethods.verified.txt
@@ -1,4 +1,4 @@
 ﻿Spectre.Console.Tests.Data.GenericException<Spectre.Console.IAnsiConsole>: Throwing!
-  at bool MethodThatThrowsGenericException<T>() in /xyz/Exceptions.cs:nn
-  at void <Should_Write_GenericException>b__8_0() in /xyz/ExceptionTests.cs:nn
-  at Exception GetException(Action action) in /xyz/ExceptionTests.cs:nn
+  at bool MethodThatThrowsGenericException<T>() in /xyz/Exceptions.cs:100      
+  at void <Should_Write_GenericException>b__8_0() in /xyz/ExceptionTests.cs:200
+  at Exception GetException(Action action) in /xyz/ExceptionTests.cs:300
