+++ src/Spectre.Console.Tests/Expectations/Exception/NoStackTrace.Output.verified.txt
@@ -1,2 +1,2 @@
 ﻿System.InvalidOperationException: Something threw!
-   System.InvalidOperationException: Throwing!
+   System.InvalidOperationException: Throwing!
