+++ src/Spectre.Console/AnsiConsole.Exceptions.cs
@@ -10,7 +10,7 @@ public static partial class AnsiConsole
     /// </summary>
     /// <param name="exception">The exception to write to the console.</param>
     /// <param name="format">The exception format options.</param>
-    [RequiresDynamicCode(ExceptionFormatter.AotWarning)]
+    [RequiresDynamicCode(ExceptionRenderableBuilder.AotWarning)]
     public static void WriteException(Exception exception, ExceptionFormats format = ExceptionFormats.Default)
     {
         Console.WriteException(exception, format);
@@ -21,7 +21,7 @@ public static partial class AnsiConsole
     /// </summary>
     /// <param name="exception">The exception to write to the console.</param>
     /// <param name="settings">The exception settings.</param>
-    [RequiresDynamicCode(ExceptionFormatter.AotWarning)]
+    [RequiresDynamicCode(ExceptionRenderableBuilder.AotWarning)]
     public static void WriteException(Exception exception, ExceptionSettings settings)
     {
         Console.WriteException(exception, settings);
