+++ src/Spectre.Console/Extensions/AnsiConsoleExtensions.Exceptions.cs
@@ -11,7 +11,7 @@ public static partial class AnsiConsoleExtensions
     /// <param name="console">The console.</param>
     /// <param name="exception">The exception to write to the console.</param>
     /// <param name="format">The exception format options.</param>
-    [RequiresDynamicCode(ExceptionFormatter.AotWarning)]
+    [RequiresDynamicCode(ExceptionRenderableBuilder.AotWarning)]
     public static void WriteException(this IAnsiConsole console, Exception exception,
         ExceptionFormats format = ExceptionFormats.Default)
     {
@@ -26,7 +26,7 @@ public static partial class AnsiConsoleExtensions
     /// <param name="console">The console.</param>
     /// <param name="exception">The exception to write to the console.</param>
     /// <param name="settings">The exception settings.</param>
-    [RequiresDynamicCode(ExceptionFormatter.AotWarning)]
+    [RequiresDynamicCode(ExceptionRenderableBuilder.AotWarning)]
     public static void WriteException(this IAnsiConsole console, Exception exception, ExceptionSettings settings)
     {
         ArgumentNullException.ThrowIfNull(console);
