+++ src/Spectre.Console/Extensions/Bcl/ExceptionExtensions.cs
@@ -11,7 +11,7 @@ public static class ExceptionExtensions
     /// <param name="exception">The exception to format.</param>
     /// <param name="format">The exception format options.</param>
     /// <returns>A <see cref="IRenderable"/> representing the exception.</returns>
-    [RequiresDynamicCode(ExceptionFormatter.AotWarning)]
+    [RequiresDynamicCode(ExceptionRenderableBuilder.AotWarning)]
     public static IRenderable GetRenderable(this Exception exception, ExceptionFormats format = ExceptionFormats.Default)
     {
         ArgumentNullException.ThrowIfNull(exception);
@@ -28,13 +28,13 @@ public static class ExceptionExtensions
     /// <param name="exception">The exception to format.</param>
     /// <param name="settings">The exception settings.</param>
     /// <returns>A <see cref="IRenderable"/> representing the exception.</returns>
-    [RequiresDynamicCode(ExceptionFormatter.AotWarning)]
+    [RequiresDynamicCode(ExceptionRenderableBuilder.AotWarning)]
     public static IRenderable GetRenderable(this Exception exception, ExceptionSettings settings)
     {
         ArgumentNullException.ThrowIfNull(exception);
 
         ArgumentNullException.ThrowIfNull(settings);
 
-        return ExceptionFormatter.Format(exception, settings);
+        return ExceptionRenderableBuilder.Format(exception, settings);
     }
 }
