+++ src/Spectre.Console/Widgets/Exceptions/ExceptionSettings.cs
@@ -8,12 +8,18 @@ public sealed class ExceptionSettings
     /// <summary>
     /// Gets or sets the exception format.
     /// </summary>
-    public ExceptionFormats Format { get; set; }
+    public ExceptionFormats Format { get; init; }
 
     /// <summary>
     /// Gets or sets the exception style.
     /// </summary>
-    public ExceptionStyle Style { get; set; }
+    public ExceptionStyle Style { get; init; }
+
+    /// <summary>
+    /// Gets or sets the exception scrubber.
+    /// Useful for removing sensitive data and for testing.
+    /// </summary>
+    public ExceptionInfoResolver Resolver { get; init; }
 
     /// <summary>
     /// Initializes a new instance of the <see cref="ExceptionSettings"/> class.
@@ -22,5 +28,6 @@ public sealed class ExceptionSettings
     {
         Format = ExceptionFormats.Default;
         Style = new ExceptionStyle();
+        Resolver = ExceptionInfoResolver.Shared;
     }
 }
